consistency - set all long fields to ints to match the data on the backend backend

+++ src/main/java/org/mythtv/services/api/dvr/DvrOperations.java
@@ -193,6 +193,6 @@ public interface DvrOperations {
 	 * @param startTime
 	 * @return
 	 */
-	ResponseEntity<Bool> removeRecorded( Long channelId, DateTime startTime ) throws MythServiceApiRuntimeException;
+	ResponseEntity<Bool> removeRecorded( int channelId, DateTime startTime ) throws MythServiceApiRuntimeException;
 
 }
